requirements.txt
torch
torchvision
albumentations
datasets
Pillow
matplotlib
numpy
scikit-learn
seaborn
tqdm
opencv-python
nbformat
albumentations
cv2
torch
dotenv
numpy
pillow
matplotlib
pandas
umap-learn
jupyter
albumentations
opencv-python
torch
torchvision
torchaudio